% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_two_stage.R
\name{des_two_stage}
\alias{des_two_stage}
\title{Design a two-stage two-arm randomised clinical trial for a binary primary
outcome variable}
\usage{
des_two_stage(type = "binomial", alpha = 0.1, beta = 0.2,
  delta = 0.2, ratio = 1, Pi0 = 0.1, Pi1 = Pi0[1], nCmax = 50L,
  equal = T, w = c(1, 0, 0, 0, 0), piO = Pi0[1], efficacy = F,
  futility = T, efficacy_type = 0L, efficacy_param = NULL,
  futility_type = 1L, futility_param = 0L, summary = F)
}
\arguments{
\item{type}{A \code{\link{character}} string indicating the chosen design
framework/test statistic to assume. Must be one of \code{"barnard"},
\code{"binomial"}, \code{"fisher"}, or \code{"single_double"}. Defaults to
\code{"binomial"}.}

\item{alpha}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&alpha;</i>}}{\eqn{\alpha}}, the significance level.
Defaults to \code{0.1}.}

\item{beta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&beta;</i>}}{\eqn{\beta}}, used in the definition of
the desired power. Defaults to \code{0.2}.}

\item{delta}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<i>&delta;</i><sub>1</sub>}}{\eqn{\delta_1}}, the
desired treatment effect. Defaults to \code{0.2}.}

\item{ratio}{A \code{\link{numeric}} indicating the chosen value for
\ifelse{html}{\out{<b><i>r</i></b>}}{\eqn{\bold{r}}}, the allocation ratio to
the experimental arm, relative to the control arm. Defaults to \code{1}.}

\item{Pi0}{A \code{\link{numeric}} \code{\link{vector}} indicating the
chosen values of the control arm response rate to allow for in the null
hypothesis. Must either be of length one, indicating a point null, or of
length two. Then, the elements indicate the range of possible response rates
to allow for. Defaults to \code{0.1}.}

\item{Pi1}{A \code{\link{numeric}} \code{\link{vector}} indicating the
chosen values of the control arm response rate to allow for in the
alternative hypothesis. Must either be of length one, indicating a point
null, or of length two. Then, the elements indicate the range of possible
response rates to allow for. Defaults to \code{Pi0[1]}.}

\item{nCmax}{A \code{\link{numeric}} indicating the maximum value of the
sample size in the control arm (across both stages) to consider. Defaults to
\code{50L}.}

\item{equal}{A \code{\link{logical}} variable indicating whether the sample
size (in each arm) in each stage should be equal. Defaults to \code{T}.}

\item{w}{A \code{\link{numeric}} \code{\link{vector}} indicating the weights
to use in the optimality criteria. Must be of length five, with all elements
greater than or equal to zero, and at least one of the first four elements
strictly positive. Defaults to \code{c(1, 0, 0, 0, 0)}.}

\item{piO}{A \code{\link{numeric}} indicating the value of the control
arm response rate to assume in the optimality criteria. Defaults to
\code{Pi0[1]}.}

\item{efficacy}{Only used if \code{type} is one of \code{"binomial"},
\code{"fisher"}, or \code{"single_double"}. Then, it is a
\code{\link{logical}} variable indicating whether to include early stopping
for efficacy in the design. Defaults to \code{F}.}

\item{futility}{Only used if \code{type} is one of \code{"binomial"},
\code{"fisher"}, or \code{"single_double"}. Then, it is a
\code{\link{logical}} variable indicating whether to include early stopping
for futility in the design. Defaults to \code{F}.}

\item{efficacy_type}{Only used if \code{type} is \code{"fisher"}. Then, it is
a \code{\link{numeric}} indicating whether, and which type of, early stopping
for efficacy to include in the design. See the vignette for details. Defaults
to \code{0L}.}

\item{efficacy_param}{Only used if \code{type} is \code{"fisher"} and
\code{efficacy_type} is not equal to \code{0L}. Then, it is a
\code{\link{numeric}} that influences the precise way in which an efficacy
boundary is specified. See the vignette for details. Defaults to \code{NULL}.}

\item{futility_type}{Only used if \code{type} is \code{"fisher"}. Then, it is
a \code{\link{numeric}} indicating whether, and which type of, early stopping
for futility to include in the design. See the vignette for details. Defaults
to \code{1L}.}

\item{futility_param}{Only used if \code{type} is \code{"fisher"} and
\code{futility_type} is not equal to \code{0L}. Then, it is a
\code{\link{numeric}} that influences the precise way in which a futility
boundary is specified. See the vignette for details. Defaults to \code{1L}.}

\item{summary}{A \code{\link{logical}} variable indicating whether a summary
of the function's progress should be printed to the console. Defaults to
\code{F}.}
}
\value{
A \code{\link{list}} of additional class \code{"ph2rand_des"},
containing each of the input parameters along with several additional
variables including:
\itemize{
\item A selection of elements that prescribe the rejection boundaries of the
optimal design. The names of these elements depends on the value of
\code{type}.
\item A \code{\link{tibble}} in the slot \code{$feasible} summarising the
operating characteristics of the feasible designs.
\item A \code{\link{numeric}} \code{\link{vector}} in the slot \code{$nC}
giving the sample sizes in the control arm in each stage for the optimal
design.
\item A \code{\link{numeric}} \code{\link{vector}} in the slot \code{$nE}
giving the sample sizes in the experimental arm in each stage for the optimal
design.
\item A \code{\link{tibble}} in the slot \code{$opchar} summarising the
operating characteristics of the optimal design.
}
}
\description{
\code{des_two_stage} determines one-stage two-arm randomised clinical trial
designs assuming the primary outcome variable is binary. More specifically,
under the assumption that it is Bernoulli distributed. It supports a flexible
framework through which the scenarios to control the type-I and type-II
error-rates for can be specified, and allows for design determination
assuming a variety of test statistics. In all instances, \code{des_two_stage}
computes the relevant required sample size in each arm, and returns
information on key operating characteristics.
}
\examples{
# The design for the default parameters
des       <- des_two_stage()
# Controlling the type-I/II error-rates over a range of possible response
rates
des_range <- des_two_stage(Pi0 = c(0, 1),
                           Pi1 = c(0, 0.8))
}
\seealso{
\code{\link{des_one_stage}}, \code{\link{pmf}},
\code{\link{terminal}}.
}
