% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ph2rand_des.R
\name{summary.ph2rand_des}
\alias{summary.ph2rand_des}
\title{Summarise a randomised clinical trial design that assumes a Bernoulli
distributed primary outcome variable}
\usage{
\method{summary}{ph2rand_des}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ph2rand_des}, as returned by
\code{\link{des_one_stage}} or \code{\link{des_two_stage}}.}

\item{...}{Not currently used.}
}
\value{
Currently not used.
}
\description{
\code{summary.ph2rand_des} prints a summary of a design returned by
\code{\link{des_one_stage}} or \code{\link{des_two_stage}}.
}
\examples{
# The default two-stage design
des <- des_two_stage()
# Print a summary
summary(des)
}
\seealso{
\code{\link{des_one_stage}}, \code{\link{des_two_stage}},
\code{\link{plot.ph2rand_des}}.
}
